preferences

+++ app/src/main/java/com/forrestguice/suntimeswidget/colors/ColorValuesEditFragment.java
@@ -1,6 +1,6 @@
 // SPDX-License-Identifier: GPL-3.0-or-later
 /*
-    Copyright (C) 2020 Forrest Guice
+    Copyright (C) 2020-2024 Forrest Guice
     This file is part of SuntimesWidget.
 
     SuntimesWidget is free software: you can redistribute it and/or modify
@@ -29,10 +29,12 @@ import android.graphics.Color;
 import android.net.Uri;
 import android.os.Build;
 import android.os.Bundle;
+import android.support.annotation.ColorRes;
 import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
 import android.support.v4.content.ContextCompat;
 import android.support.v7.app.AlertDialog;
+import android.text.SpannableString;
 import android.util.Log;
 import android.util.TypedValue;
 import android.view.Gravity;
@@ -45,12 +47,17 @@ import android.view.ViewGroup;
 import android.widget.EditText;
 import android.widget.GridLayout;
 import android.widget.ImageButton;
+import android.widget.LinearLayout;
 import android.widget.PopupMenu;
 import android.widget.TextView;
 
 import com.forrestguice.suntimeswidget.R;
+import com.forrestguice.suntimeswidget.SuntimesUtils;
+import com.forrestguice.suntimeswidget.settings.AppSettings;
 import com.forrestguice.suntimeswidget.settings.colors.ColorActivity;
 import com.forrestguice.suntimeswidget.settings.colors.ColorDialog;
+import com.forrestguice.suntimeswidget.settings.colors.ColorUtils;
+import com.forrestguice.suntimeswidget.settings.colors.pickers.ColorPickerFragment;
 import com.forrestguice.suntimeswidget.views.ViewUtils;
 
 import java.util.ArrayList;
@@ -64,7 +71,10 @@ public class ColorValuesEditFragment extends ColorValuesFragment
     public static final String ARG_ALLOW_DELETE = "allowDelete";
     public static final boolean DEF_ALLOW_DELETE = true;
 
-    protected EditText editID;
+    public static final String ARG_SHOW_ALPHA = "showAlpha";
+    public static final boolean DEF_SHOW_ALPHA = false;
+
+    protected EditText editID, editLabel;
     protected GridLayout panel;
     protected ImageButton cancelButton;
 
@@ -103,7 +113,9 @@ public class ColorValuesEditFragment extends ColorValuesFragment
 
         panel = (GridLayout) content.findViewById(R.id.colorPanel);
         editID = (EditText) content.findViewById(R.id.editTextID);
+        editLabel = (EditText) content.findViewById(R.id.editTextLabel);
         setID(null);
+        setLabel(null);
 
         if (savedState != null) {
             onRestoreInstanceState(savedState);
@@ -133,6 +145,20 @@ public class ColorValuesEditFragment extends ColorValuesFragment
         }
     }
 
+    /**
+     * @param colorsLabel value to set on edittext; use null to get the label from ColorValues
+     */
+    protected void setLabel(@Nullable String colorsLabel)
+    {
+        if (editLabel != null)
+        {
+            if (colorValues != null && colorsLabel == null) {
+                colorsLabel = colorValues.getLabel();
+            }
+            editLabel.setText(colorsLabel != null ? colorsLabel : "");
+        }
+    }
+
     protected boolean validateInput()
     {
         String colorsID = editID.getText().toString();
@@ -144,8 +170,15 @@ public class ColorValuesEditFragment extends ColorValuesFragment
             editID.setError(getString(R.string.error_colorid_spaces));
             editID.setSelection(colorsID.indexOf(" "), colorsID.indexOf(" ") + 1);
             return false;
+        }
 
-        } else return true;
+        String colorsLabel = editLabel.getText().toString();
+        if (colorsLabel.trim().isEmpty()) {
+            editLabel.setError(getString(R.string.error_colorlabel_empty));
+            return false;
+        }
+
+        return true;
     }
 
     private View.OnClickListener onSaveButtonClicked = new View.OnClickListener() {
@@ -154,17 +187,21 @@ public class ColorValuesEditFragment extends ColorValuesFragment
             onSaveColorValues();
         }
     };
-    protected void onSaveColorValues()
+    protected boolean onSaveColorValues()
     {
         if (validateInput())
         {
             String colorsID = editID.getText().toString();
+            String colorsLabel = editLabel.getText().toString();
             colorValues.setID(colorsID);
+            colorValues.setLabel(colorsLabel);
 
             if (listener != null) {
                 listener.onSaveClicked(colorsID, colorValues);
             }
+            return true;
         }
+        return false;
     }
 
     private View.OnClickListener onCancelButtonClicked = new View.OnClickListener() {
@@ -203,6 +240,7 @@ public class ColorValuesEditFragment extends ColorValuesFragment
         out.putParcelable("defaultValues", defaultValues);
         out.putStringArray("filterValues", filterValues.toArray(new String[0]));
         out.putString("editID", editID.getText().toString());
+        out.putString("editLabel", editLabel.getText().toString());
         super.onSaveInstanceState(out);
     }
     protected void onRestoreInstanceState(@NonNull Bundle savedState)
@@ -214,6 +252,7 @@ public class ColorValuesEditFragment extends ColorValuesFragment
             filterValues  = new TreeSet<>(Arrays.asList(filter));
         }
         setID(savedState.getString("editID"));
+        setLabel(savedState.getString("editLabel"));
     }
 
     @Override
@@ -229,7 +268,6 @@ public class ColorValuesEditFragment extends ColorValuesFragment
         }
     }
 
-    //protected TextView[] colorEdits;
     protected void updateViews()
     {
         if (panel != null && colorValues != null)
@@ -247,6 +285,9 @@ public class ColorValuesEditFragment extends ColorValuesFragment
                 itemMargin = (int)context.getResources().getDimension(R.dimen.colortext_margin);
             }
 
+            int[] defaultColors = getDefaultColors(context);
+            float textSizePx = getTextSizePx_medium(context);
+
             int c = 0;
             panel.removeAllViews();
             for (int i=0; i<keys.length; i++)
@@ -256,10 +297,53 @@ public class ColorValuesEditFragment extends ColorValuesFragment
                     continue;
                 }
 
+                boolean bold = true;
+                Integer textColor;
+                Integer backgroundColor;
+                switch (colorValues.getRole(keys[i]))
+                {
+                    case ColorValues.ROLE_BACKGROUND_PRIMARY: case ColorValues.ROLE_BACKGROUND:
+                    case ColorValues.ROLE_BACKGROUND_INVERSE: case ColorValues.ROLE_ACTION:
+                        bold = false;
+                        backgroundColor = colorValues.getColor(keys[i]);
+                        textColor = getContrastingColor(keys[i], ColorUtils.isTextReadable(defaultColors[0], backgroundColor)
+                                ? defaultColors[0] : defaultColors[2]);
+                        break;
+
+                    case ColorValues.ROLE_TEXT: case ColorValues.ROLE_TEXT_PRIMARY:
+                    case ColorValues.ROLE_TEXT_PRIMARY_INVERSE: case ColorValues.ROLE_TEXT_INVERSE:
+                    case ColorValues.ROLE_ACCENT: case ColorValues.ROLE_FOREGROUND:
+                        textColor = colorValues.getColor(keys[i]);
+                        backgroundColor = getContrastingColor(keys[i], defaultColors[1]);
+                        break;
+
+                    case ColorValues.ROLE_UNKNOWN:
+                    default:
+                        textColor = colorValues.getColor(keys[i]);
+                        backgroundColor = null;
+                        break;
+                }
+
+                SpannableString colorLabel = null;
+                String labelText = " " + colorValues.getLabel(keys[i]) + " ";
+
+                if (backgroundColor != null && textColor != null)
+                {
+                    float cornerRadiusPx = context.getResources().getDimension(R.dimen.chip_radius);
+                    colorLabel = SuntimesUtils.createRoundedBackgroundColorSpan(colorLabel, " " + labelText + " ", labelText, textColor, bold, backgroundColor, cornerRadiusPx, cornerRadiusPx);
+
+                } else if (textColor != null) {
+                    colorLabel = (bold ? SuntimesUtils.createBoldColorSpan(colorLabel, labelText, labelText, textColor)
+                                       : SuntimesUtils.createColorSpan(colorLabel, labelText, labelText, textColor));
+
+                } else {
+                    colorLabel = new SpannableString(" ");
+                }
+
                 TextView colorEdit = new TextView(getActivity());
-                colorEdit.setText(colorValues.getLabel(keys[i]));
-                colorEdit.setTextColor(colorValues.getColor(keys[i]));
+                colorEdit.setText(colorLabel);
                 colorEdit.setOnClickListener(onColorEditClick(keys[i]));
+                colorEdit.setTextSize(TypedValue.COMPLEX_UNIT_PX, textSizePx);
                 colorEdit.setGravity(Gravity.CENTER);
 
                 if (itemBackground != null) {
@@ -305,6 +389,7 @@ public class ColorValuesEditFragment extends ColorValuesFragment
     {
         colorValues = v;
         setID(null);
+        setLabel(null);
         updateViews();
     }
     public ColorValues getColorValues() {
@@ -369,22 +454,95 @@ public class ColorValuesEditFragment extends ColorValuesFragment
         }
     }
 
+    protected float getTextSizePx_medium(Context context)
+    {
+        int[] attr = { R.attr.text_size_medium };
+        TypedArray typedArray = context.obtainStyledAttributes(attr);
+        float textSize = typedArray.getDimension(0, context.getResources().getDimension(R.dimen.text_size_medium));
+        typedArray.recycle();
+        return textSize;
+    }
+
+    @SuppressLint("ResourceType")
+    protected int[] getDefaultColors(Context context)
+    {
+        int[] attr = { R.attr.timeCardBackground, R.attr.text_primaryColor, R.attr.text_primaryInverseColor };
+        TypedArray typedArray = context.obtainStyledAttributes(attr);
+        int backgroundColor = ContextCompat.getColor(context, typedArray.getResourceId(0, R.color.card_bg));
+        int textColor = ContextCompat.getColor(context, typedArray.getResourceId(1, R.color.text_primary));
+        int inverseTextColor = ContextCompat.getColor(context, typedArray.getResourceId(2, R.color.text_primary_inverse));
+        typedArray.recycle();
+        return new int[] { textColor, backgroundColor, inverseTextColor };
+    }
+
+    protected int[] getColorOverUnder(Context context, String key)
+    {
+        int[] defaultColors = getDefaultColors(context);
+        int colorOver = defaultColors[0];
+        int colorUnder = defaultColors[1];
+        Integer contrastingColor = getContrastingColor(key, null);
+
+        switch (colorValues.getRole(key))
+        {
+            case ColorValues.ROLE_BACKGROUND_PRIMARY:
+                colorOver = (contrastingColor != null) ? contrastingColor : defaultColors[0];
+                break;
+
+            case ColorValues.ROLE_FOREGROUND:
+            case ColorValues.ROLE_TEXT: case ColorValues.ROLE_TEXT_INVERSE:
+            case ColorValues.ROLE_TEXT_PRIMARY: case ColorValues.ROLE_TEXT_PRIMARY_INVERSE:
+                colorUnder = (contrastingColor != null) ? contrastingColor : defaultColors[1];
+                break;
+        }
+        return new int[] { colorOver, colorUnder };
+    }
+
+    @Nullable
+    protected Integer getContrastingColor(String key, Integer defaultValue)
+    {
+        String k;
+        switch (colorValues.getRole(key))
+        {
+            case ColorValues.ROLE_ACTION:
+            case ColorValues.ROLE_BACKGROUND: case ColorValues.ROLE_BACKGROUND_PRIMARY:
+                k = colorValues.findColorWithRole(ColorValues.ROLE_TEXT_PRIMARY);
+                return (k != null) ? Integer.valueOf(colorValues.getColor(k)) : defaultValue;
+
+            case ColorValues.ROLE_BACKGROUND_INVERSE:
+                k = colorValues.findColorWithRole(ColorValues.ROLE_TEXT_PRIMARY_INVERSE);
+                return (k != null) ? Integer.valueOf(colorValues.getColor(k)) : defaultValue;
+
+            case ColorValues.ROLE_TEXT: case ColorValues.ROLE_TEXT_PRIMARY:
+            case ColorValues.ROLE_ACCENT: case ColorValues.ROLE_FOREGROUND:
+                k = colorValues.findColorWithRole(ColorValues.ROLE_BACKGROUND_PRIMARY);
+                return (k != null) ? Integer.valueOf(colorValues.getColor(k)) : defaultValue;
+
+            case ColorValues.ROLE_TEXT_INVERSE:
+            case ColorValues.ROLE_TEXT_PRIMARY_INVERSE:
+                k = colorValues.findColorWithRole(ColorValues.ROLE_BACKGROUND_INVERSE);
+                return (k != null) ? Integer.valueOf(colorValues.getColor(k)) : defaultValue;
+
+            default:
+                return null;
+        }
+    }
+
     protected Intent pickColorIntent(String key, int requestCode)
     {
-        Context context = getActivity();
-        int[] attr = { R.attr.timeCardBackground, R.attr.text_primaryColor };
-        TypedArray typedArray = context.obtainStyledAttributes(attr);
-        int colorUnder = ContextCompat.getColor(context, typedArray.getResourceId(0, R.color.card_bg));
-        @SuppressLint("ResourceType")
-        int colorOver = ContextCompat.getColor(context, typedArray.getResourceId(1, R.color.text_primary));
-        typedArray.recycle();
+        int color = colorValues.getColor(key);
+        ArrayList<Integer> recentColors = new ArrayList<>(new LinkedHashSet<>(colorValues.getColors()));
+        recentColors.add(0, color);
 
         Intent intent = new Intent(getActivity(), ColorActivity.class);
-        intent.putExtra(ColorDialog.KEY_SHOWALPHA, true);
-        intent.setData(Uri.parse("color://" + String.format("#%08X", colorValues.getColor(key))));
-        intent.putExtra(ColorDialog.KEY_COLOR_UNDER, colorUnder);
-        intent.putExtra(ColorDialog.KEY_COLOR_OVER, colorOver);
-        intent.putExtra(ColorDialog.KEY_RECENT, new ArrayList<>(new LinkedHashSet<>(colorValues.getColors())));
+        intent.putExtra(ColorDialog.KEY_SHOWALPHA, showAlpha());
+        intent.setData(Uri.parse("color://" + String.format("#%08X", color)));
+        intent.putExtra(ColorDialog.KEY_RECENT, recentColors);
+        intent.putExtra(ColorDialog.KEY_LABEL, colorValues.getLabel(key));
+
+        int[] colorOverUnder = getColorOverUnder(getActivity(), key);
+        intent.putExtra(ColorDialog.KEY_COLOR_OVER, colorOverUnder[0]);
+        intent.putExtra(ColorDialog.KEY_COLOR_UNDER, colorOverUnder[1]);
+        intent.putExtra(ColorDialog.KEY_PREVIEW_MODE, ColorPickerFragment.ColorPickerModel.PREVIEW_LUMINANCE);
 
         if (defaultValues != null) {
             intent.putExtra(ColorDialog.KEY_SUGGESTED, defaultValues.getColor(key));
@@ -443,15 +601,18 @@ public class ColorValuesEditFragment extends ColorValuesFragment
 
     protected void shareColors(Context context)
     {
-        Intent intent = new Intent(Intent.ACTION_SEND);
-        intent.setType("text/plain");
-        intent.putExtra(Intent.EXTRA_TEXT, colorValues.toString());
-        startActivity(Intent.createChooser(intent, null));
+        if (colorValues != null)
+        {
+            Intent intent = new Intent(Intent.ACTION_SEND);
+            intent.setType("text/plain");
+            intent.putExtra(Intent.EXTRA_TEXT, colorValues.toString());
+            startActivity(Intent.createChooser(intent, null));
+        }
     }
 
     protected void deleteColors(Context context)
     {
-        if (listener != null && allowDelete()) {
+        if (listener != null && allowDelete() && colorValues != null) {
             listener.onDeleteClicked(colorValues.getID());
         }
     }
@@ -460,6 +621,13 @@ public class ColorValuesEditFragment extends ColorValuesFragment
     }
     public void setAllowDelete(boolean allowDelete) {
         setBoolArg(ARG_ALLOW_DELETE, allowDelete);
+    }
+
+    public boolean showAlpha() {
+        return getBoolArg(ARG_SHOW_ALPHA, DEF_SHOW_ALPHA);
+    }
+    public void setShowAlpha(boolean value) {
+        setBoolArg(ARG_SHOW_ALPHA, value);
     }
 
     public static final int REQUEST_IMPORT_THEME = 1000;
@@ -489,7 +657,7 @@ public class ColorValuesEditFragment extends ColorValuesFragment
         }
     }
 
-    private PopupMenu.OnMenuItemClickListener onOverflowMenuItemSelected = new PopupMenu.OnMenuItemClickListener()
+    private final PopupMenu.OnMenuItemClickListener onOverflowMenuItemSelected = new PopupMenu.OnMenuItemClickListener()
     {
         @Override
         public boolean onMenuItemClick(MenuItem item)
@@ -519,9 +687,11 @@ public class ColorValuesEditFragment extends ColorValuesFragment
     @Override
     public boolean onOptionsItemSelected(MenuItem item)
     {
-        if (onOverflowMenuItemSelected.onMenuItemClick(item)) {
-            return true;
-        } else { return super.onOptionsItemSelected(item); }
+        //if (onOverflowMenuItemSelected.onMenuItemClick(item)) {
+        //    return true;
+        //} else {
+            return super.onOptionsItemSelected(item);
+        //}
     }
 
     protected void setBoolArg(String key, boolean value) {
